+++ NEWS.md
@@ -5,6 +5,7 @@
 - add support for formula in getchain to find compounds ions
 - remove frequency cutoff in getsda and use PMD network clusters analysis to determine the cutoff
 - add merge feature for getcluster and such methods could be used to furthor reduce the GlobalStd peaks
+- remove hmdbp data since pmd network analysis could cover this topic
 
 # pmd 0.1.5